refactor: split client into binary and lib
