+++ cargo-shuttle/src/main.rs
@@ -1,221 +1,8 @@
-mod args;
-mod client;
-mod config;
-mod factory;
-mod print;
-
-use std::fs::File;
-use std::io::Write;
-use std::io::{self, stdout};
-use std::net::{Ipv4Addr, SocketAddr};
-use std::rc::Rc;
-use std::sync::mpsc;
-
-use anyhow::{Context, Result};
-use args::{LoginArgs, ProjectArgs};
-use cargo::core::resolver::CliFeatures;
-use cargo::core::Workspace;
-use cargo::ops::{PackageOpts, Packages};
-use factory::LocalFactory;
-use futures::future::TryFutureExt;
-use shuttle_service::loader::{build_crate, Loader};
+use anyhow::Result;
+use cargo_shuttle::{Args, Shuttle};
 use structopt::StructOpt;
-use tokio::task::spawn_blocking;
-use uuid::Uuid;
-
-use crate::args::{Args, AuthArgs, Command, DeployArgs};
-use crate::config::RequestContext;
 
 #[tokio::main]
 async fn main() -> Result<()> {
     Shuttle::new().run(Args::from_args()).await
-}
-
-pub struct Shuttle {
-    ctx: RequestContext,
-}
-
-impl Default for Shuttle {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl Shuttle {
-    pub fn new() -> Self {
-        let ctx = RequestContext::load_global().unwrap();
-        Self { ctx }
-    }
-
-    pub async fn run(mut self, args: Args) -> Result<()> {
-        if matches!(
-            args.cmd,
-            Command::Deploy(..) | Command::Delete | Command::Status | Command::Logs | Command::Run
-        ) {
-            self.load_project(&args.project_args)?;
-        }
-
-        self.ctx.set_api_url(args.api_url);
-
-        match args.cmd {
-            Command::Deploy(deploy_args) => self.deploy(deploy_args).await,
-            Command::Status => self.status().await,
-            Command::Logs => self.logs().await,
-            Command::Delete => self.delete().await,
-            Command::Auth(auth_args) => self.auth(auth_args).await,
-            Command::Login(login_args) => self.login(login_args).await,
-            Command::Run => self.local_run().await,
-        }
-    }
-
-    pub fn load_project(&mut self, project_args: &ProjectArgs) -> Result<()> {
-        self.ctx.load_local(project_args)
-    }
-
-    async fn login(&mut self, login_args: LoginArgs) -> Result<()> {
-        let api_key_str = login_args.api_key.unwrap_or_else(|| {
-            let url = "https://shuttle.rs/login";
-
-            let _ = webbrowser::open(url);
-
-            println!("If your browser did not automatically open, go to {url}");
-            print!("Enter Api Key: ");
-
-            io::stdout().flush().unwrap();
-
-            let mut input = String::new();
-
-            io::stdin().read_line(&mut input).unwrap();
-
-            input
-        });
-
-        let api_key = api_key_str.trim().parse()?;
-
-        self.ctx.set_api_key(api_key)?;
-
-        Ok(())
-    }
-
-    async fn auth(&mut self, auth_args: AuthArgs) -> Result<()> {
-        let api_key = client::auth(self.ctx.api_url(), auth_args.username)
-            .await
-            .context("failed to retrieve api key")?;
-        self.ctx.set_api_key(api_key)?;
-        Ok(())
-    }
-
-    async fn delete(&self) -> Result<()> {
-        client::delete(
-            self.ctx.api_url(),
-            self.ctx.api_key()?,
-            self.ctx.project_name(),
-        )
-        .await
-        .context("failed to delete deployment")
-    }
-
-    async fn status(&self) -> Result<()> {
-        client::status(
-            self.ctx.api_url(),
-            self.ctx.api_key()?,
-            self.ctx.project_name(),
-        )
-        .await
-        .context("failed to get status of deployment")
-    }
-
-    async fn logs(&self) -> Result<()> {
-        client::logs(
-            self.ctx.api_url(),
-            self.ctx.api_key()?,
-            self.ctx.project_name(),
-        )
-        .await
-        .context("failed to get logs of deployment")
-    }
-
-    async fn local_run(&self) -> Result<()> {
-        let buf = Box::new(stdout());
-        let working_directory = self.ctx.working_directory();
-        let so_path = build_crate(working_directory, buf)?;
-        let loader = Loader::from_so_file(so_path)?;
-
-        let mut factory = LocalFactory {};
-        let addr = SocketAddr::new(Ipv4Addr::LOCALHOST.into(), 8000);
-        let deployment_id = Uuid::new_v4();
-        let (tx, rx) = mpsc::sync_channel(100);
-
-        let (handle, so) = loader.load(&mut factory, addr, tx, deployment_id)?;
-
-        let logs = spawn_blocking(move || loop {
-            while let Ok(log) = rx.recv() {
-                print::log(log.datetime, log.item);
-            }
-        });
-
-        handle.await??;
-
-        tokio::spawn(async move {
-            so.close().unwrap();
-        });
-
-        logs.await?;
-
-        Ok(())
-    }
-
-    async fn deploy(&self, args: DeployArgs) -> Result<()> {
-        let package_file = self
-            .run_cargo_package(args.allow_dirty)
-            .context("failed to package cargo project")?;
-
-        let key = self.ctx.api_key()?;
-
-        client::deploy(
-            package_file,
-            self.ctx.api_url(),
-            key,
-            self.ctx.project_name(),
-        )
-        .and_then(|_| {
-            client::secrets(
-                self.ctx.api_url(),
-                key,
-                self.ctx.project_name(),
-                self.ctx.secrets(),
-            )
-        })
-        .await
-        .context("failed to deploy cargo project")
-    }
-
-    // Packages the cargo project and returns a File to that file
-    fn run_cargo_package(&self, allow_dirty: bool) -> Result<File> {
-        let config = cargo::util::config::Config::default()?;
-
-        let working_directory = self.ctx.working_directory();
-        let path = working_directory.join("Cargo.toml");
-
-        let ws = Workspace::new(&path, &config)?;
-        let opts = PackageOpts {
-            config: &config,
-            list: false,
-            check_metadata: true,
-            allow_dirty,
-            verify: false,
-            jobs: None,
-            to_package: Packages::Default,
-            targets: vec![],
-            cli_features: CliFeatures {
-                features: Rc::new(Default::default()),
-                all_features: false,
-                uses_default_features: true,
-            },
-        };
-
-        let locks = cargo::ops::package(&ws, &opts)?.expect("unwrap ok here");
-        let owned = locks.get(0).unwrap().file().try_clone()?;
-        Ok(owned)
-    }
 }
